backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-silk==5.3.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gprof2dot==2024.6.6
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pluggy==1.5.0
pycodestyle==2.13.0
PyJWT==2.9.0
pytest==8.3.5
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.1
uritemplate==4.1.1
